Remove UUID import, check date using regex, update error handling
