+++ package.json
@@ -35,8 +35,7 @@
     "typedi": "^0.10.0",
     "typescript": "^4.4.4",
     "winston": "^3.3.3",
-    "winston-daily-rotate-file": "^4.5.5",
-    "uuid": "^8.3.2"
+    "winston-daily-rotate-file": "^4.5.5"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
@@ -51,7 +50,6 @@
     "eslint": "^8.1.0",
     "eslint-config-airbnb-typescript": "^14.0.1",
     "eslint-plugin-import": "^2.25.2",
-    "nodemon": "^2.0.14",
-    "@types/uuid": "^8.3.0"
+    "nodemon": "^2.0.14"
   }
 }
